# examples/openapi-tools/quickstart-se/se-server/src/main/resources/META-INF/openapi.yml
#
# Copyright (c) 2022, 2024 Oracle and/or its affiliates.
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#     http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.
#
openapi: 3.0.0
info:
  description: This is a sample for Helidon Quickstart project.
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  title: OpenAPI Helidon Quickstart
  version: 1.0.0
servers:
- url: http://localhost:8080
tags:
- name: message
paths:
  /greet:
    get:
      operationId: getDefaultMessage
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
          description: successful operation
      summary: Return a worldly greeting message.
      tags:
      - message
      x-accepts: application/json
  /greet/greeting:
    put:
      operationId: updateGreeting
      requestBody:
        $ref: '#/components/requestBodies/Message'
      responses:
        "200":
          description: successful operation
        "400":
          description: No greeting provided
      summary: Set the greeting to use in future messages.
      tags:
      - message
      x-content-type: application/json
      x-accepts: application/json
  /greet/{name}:
    get:
      operationId: getMessage
      parameters:
      - description: the name to greet
        explode: false
        in: path
        name: name
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
          description: successful operation
      summary: Return a greeting message using the name that was provided.
      tags:
      - message
      x-accepts: application/json
components:
  requestBodies:
    Message:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Message'
      description: Message for the user
      required: true
  schemas:
    Message:
      description: An message for the user
      example:
        greeting: greeting
        message: message
      properties:
        message:
          format: int64
          type: string
        greeting:
          format: int64
          type: string
      type: object